Deprecate ftn_spname option and update related documentation
Updated the ftn_spname option to be deprecated in favor of allow_authorize_params for adding query parameters. Enhanced documentation to reflect these changes and added warnings in the logger for deprecated methods. Removed the load_signing_key method for cleaner code and future-proofing. Additionally, added a filter for version.rb in test coverage.