Update CHANGELOG for version 1.3.1 and adjust Gemfile.lock and version.rb
This commit updates the CHANGELOG to reflect the new version 1.3.1, highlighting enhancements to SSL configuration, updates to the federation controller, route adjustments, gemfile modifications for Rails 8 compatibility, and improvements in time handling. Additionally, it updates the Gemfile.lock to specify the new omniauth_openid_federation version and modifies version.rb to reflect the version change.